RT #74224: loaded() noisy under Perl 5.15.7.
The immediate problem here is that 'defined(@array)' has been deprecated
for a while, but only with Perl 5.15.7 does Perl complain when @array is
a package variable.

The next problem is how to change the code. The standard response for
Boolean context is just to say '@array'. This is what the patch does,
but it was not straightforward to arrive at this. Here's the deal:

The original code was 'defined @{"${name}::ISA"}. The question is
whether this tests whether the array is occupied, or whether it tests
whether the array has been created, independant of whether it is
occupied.

A little playing around with 'perl -E' shows that '@{"Foo::Bar"}'
autovivifies array Bar in package Foo. That is,
  $ perl -E '@{"Foo::Bar"}; say *Foo::Bar{ARRAY}'
prints an array reference, while
  $ perl -E 'say *Foo::Bar{ARRAY}'
does not. So I was tempted to recast the code in terms of the glob.

But
 $ perl -E 'say *Foo::ISA{ARRAY}'
prints an array reference, so it appears the @ISA already exists, so
that '*"${name}::ISA"{ARRAY}' would always return a true value. This
seems to leave us with the occupancy test as the only alternative.

But tricky Perl is not my strong suit, and you may well find a better
implementation. The patch is just to try to get the ball rolling.